implements pr feedback
